+++ src/features/UserPosts/type.ts
